Refactor routes to separate files
