+++ server/routes/primary.js
@@ -2,10 +2,6 @@ const dbModels = require('../../db/index.js');
 const utils = require('../utils.js');
 const url = require('url');
 
-const cloudinary = require('cloudinary');
-const cloudinaryConfig = require('../config/cloudinary.js');
-cloudinary.config(cloudinaryConfig.info);
-
 module.exports = function(app){
   app.get('/', function(req, res, next) {
     res.redirect('/index.html');
@@ -15,7 +11,7 @@ module.exports = function(app){
     res.redirect('/createEvent.html');
   });
 
-  // post new user
+  // New user
   app.post('/user', function(req, res, next) {
     dbModels.User
     .create({
@@ -144,70 +140,6 @@ module.exports = function(app){
     .catch(function(error) {
       console.log('POST Error', error);
       next(error);
-    });
-  });
-
-  app.get('/itemList', function(req, res, next) {
-    var eventName = url.parse(req.url).query.slice(10).split('_').join(' ');
-    dbModels.Event.findOne({where: {name: eventName}})
-    .then(function(event) {
-      var eventId = event.id;
-      dbModels.ItemList.findAll({where: {eventId: eventId}})
-      .then(function(items) {
-        utils.sendResponse(res, 200, 'application/json', items);
-      });
-    });
-  });
-
-  app.post('/itemList', function(req, res, next) {
-    var eventName = url.parse(req.url).query.slice(10).split('_').join(' ');
-    dbModels.Event.findOne({where: {name: eventName}})
-    .then(function(event) {
-      var eventId = event.id;
-      dbModels.ItemList
-      .create({
-        item: req.body.item,
-        owner: req.body.owner,
-        cost: req.body.cost,
-        eventId: eventId
-      })
-      .then(function(item) {
-        utils.sendResponse(res, 201, 'text/html', 'item successfully posted');
-      }).catch(function(err) {
-        console.log('Error: ', err);
-      });
-    });
-  });
-
-  app.get('/reminders', function(req, res, next) {
-    var eventName = url.parse(req.url).query.slice(10).split('_').join(' ');
-    dbModels.Event.findOne({where: {name: eventName}})
-    .then(function(event) {
-      var eventId = event.id;
-      dbModels.Reminder.findAll({where: {eventId: eventId}})
-      .then(function(reminders) {
-        utils.sendResponse(res, 200, 'application/json', reminders);
-      });
-    });
-  });
-  
-  app.post('/reminders', function(req, res, next) {
-    var eventName = url.parse(req.url).query.slice(10).split('_').join(' ');
-    dbModels.Event.findOne({where: {name: eventName}})
-    .then(function(event) {
-      var eventId = event.id;
-      dbModels.Reminder
-      .create({
-        phoneNumber: req.body.phoneNumber,
-        msg: req.body.msg,
-        when: req.body.when,
-        eventId: eventId
-      })
-      .then(function(item) {
-        utils.sendResponse(res, 201, 'text/html', 'reminder successfully posted');
-      }).catch(function(err) {
-        console.log('Error: ', err);
-      });
     });
   });
 }
